Split compilation
The Initialize_Control_Board function was declared in game_setting.h
and modified to be defined in game_setting.c.

resolved : #14
