+++ game_setting.h
@@ -1,0 +1,4 @@
+//The game board can change by changing this defined constant
+#define BOARD_SIZE 6
+
+int Initialize_Control_Board(char control_board[BOARD_SIZE][BOARD_SIZE], int difficulty);
